fix: pisahkan layout per route group untuk hilangkan double sidebar (#63)
- (dashboard)/layout.tsx diubah jadi pure auth wrapper tanpa UI
- Buat superadmin/layout.tsx dengan sidebar+header sendiri
- Portal page (dashboard)/page.tsx embed Header langsung karena parent layout tidak lagi render UI
