+++ app/(dashboard)/page.tsx
@@ -6,6 +6,7 @@ import type { UserRole } from '@/lib/auth-helpers'
 import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardFooter, CardHeader } from '@/components/ui/card'
+import { Header } from '@/components/layout/header'
 import type { Subapp } from '@/lib/db/schema'
 
 const TYPE_LABEL: Record<string, string> = {
@@ -67,63 +68,76 @@ export default async function DashboardPortalPage() {
   }
 
   return (
-    <div className="flex flex-col gap-6">
-      <div>
-        <h1 className="text-2xl font-semibold tracking-tight">
-          Selamat Datang, {user.name}
-        </h1>
-        <p className="text-muted-foreground mt-1">
-          Pilih aplikasi yang ingin Anda akses.
-        </p>
-      </div>
+    <div className="flex min-h-screen flex-col">
+      <Header
+        userName={user.name}
+        userEmail={user.email}
+        userAvatar={(user as { avatar?: string | null }).avatar ?? null}
+        userRole={userRole}
+        subapps={subappList.filter((s) => s.type !== 'superadmin')}
+        currentSubappKey={null}
+      />
 
-      {subappList.length === 0 ? (
-        <p className="text-muted-foreground">
-          Anda belum memiliki akses ke sub-aplikasi manapun. Hubungi administrator.
-        </p>
-      ) : (
-        <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
-          {subappList.map((subapp) => (
-            <Card key={subapp.id} className="flex flex-col">
-              <CardHeader className="items-center pb-2">
-                {subapp.image ? (
-                  <div className="relative size-16 overflow-hidden rounded-full border">
-                    <Image
-                      src={subapp.image}
-                      alt={subapp.name ?? subapp.key}
-                      fill
-                      className="object-cover"
-                    />
-                  </div>
-                ) : (
-                  <div className="flex size-16 items-center justify-center rounded-full border bg-muted text-2xl font-bold text-muted-foreground">
-                    {(subapp.name ?? subapp.key).slice(0, 1).toUpperCase()}
-                  </div>
-                )}
-              </CardHeader>
+      <main className="flex-1 overflow-y-auto p-4 lg:p-6">
+        <div className="flex flex-col gap-6">
+          <div>
+            <h1 className="text-2xl font-semibold tracking-tight">
+              Selamat Datang, {user.name}
+            </h1>
+            <p className="text-muted-foreground mt-1">
+              Pilih aplikasi yang ingin Anda akses.
+            </p>
+          </div>
 
-              <CardContent className="flex flex-1 flex-col items-center gap-2 text-center">
-                <p className="font-medium leading-tight">
-                  {subapp.name ?? subapp.key}
-                </p>
-                <Badge variant={TYPE_VARIANT[subapp.type] ?? 'outline'}>
-                  {TYPE_LABEL[subapp.type] ?? subapp.type}
-                </Badge>
-              </CardContent>
+          {subappList.length === 0 ? (
+            <p className="text-muted-foreground">
+              Anda belum memiliki akses ke sub-aplikasi manapun. Hubungi administrator.
+            </p>
+          ) : (
+            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
+              {subappList.map((subapp) => (
+                <Card key={subapp.id} className="flex flex-col">
+                  <CardHeader className="items-center pb-2">
+                    {subapp.image ? (
+                      <div className="relative size-16 overflow-hidden rounded-full border">
+                        <Image
+                          src={subapp.image}
+                          alt={subapp.name ?? subapp.key}
+                          fill
+                          className="object-cover"
+                        />
+                      </div>
+                    ) : (
+                      <div className="flex size-16 items-center justify-center rounded-full border bg-muted text-2xl font-bold text-muted-foreground">
+                        {(subapp.name ?? subapp.key).slice(0, 1).toUpperCase()}
+                      </div>
+                    )}
+                  </CardHeader>
 
-              <CardFooter className="justify-center pt-2">
-                <Button
-                  size="sm"
-                  className="w-full"
-                  render={<Link href={subappHref(subapp)} />}
-                >
-                  Masuk
-                </Button>
-              </CardFooter>
-            </Card>
-          ))}
+                  <CardContent className="flex flex-1 flex-col items-center gap-2 text-center">
+                    <p className="font-medium leading-tight">
+                      {subapp.name ?? subapp.key}
+                    </p>
+                    <Badge variant={TYPE_VARIANT[subapp.type] ?? 'outline'}>
+                      {TYPE_LABEL[subapp.type] ?? subapp.type}
+                    </Badge>
+                  </CardContent>
+
+                  <CardFooter className="justify-center pt-2">
+                    <Button
+                      size="sm"
+                      className="w-full"
+                      render={<Link href={subappHref(subapp)} />}
+                    >
+                      Masuk
+                    </Button>
+                  </CardFooter>
+                </Card>
+              ))}
+            </div>
+          )}
         </div>
-      )}
+      </main>
     </div>
   )
 }
